+++ dropwizard-v5/src/test/scala/fr/davit/akka/http/metrics/dropwizard/DropwizardRegistrySpec.scala
@@ -16,20 +16,17 @@
 
 package fr.davit.akka.http.metrics.dropwizard
 
-import akka.http.scaladsl.model.StatusCodes
-import fr.davit.akka.http.metrics.core.Dimension
-import fr.davit.akka.http.metrics.core.HttpMetricsRegistry.{PathDimension, StatusGroupDimension}
+import fr.davit.akka.http.metrics.core.{Dimension, PathLabeler, StatusGroupLabeler}
 import io.dropwizard.metrics5.MetricName
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 
 import scala.concurrent.duration._
-
 import scala.jdk.CollectionConverters._
 
 class DropwizardRegistrySpec extends AnyFlatSpec with Matchers {
 
-  val dimensions = Seq(StatusGroupDimension(StatusCodes.OK), PathDimension("/api"))
+  val dimensions = Seq(Dimension(StatusGroupLabeler.name, "2xx"), Dimension(PathLabeler.name, "/api"))
 
   trait Fixture {
     val registry = DropwizardRegistry()
@@ -47,7 +44,7 @@ class DropwizardRegistrySpec extends AnyFlatSpec with Matchers {
     }
 
     private def metricName(name: String, dims: Seq[Dimension]): MetricName = {
-      MetricName.build(name).tagged(dims.map(d => d.key -> d.value).toMap.asJava)
+      MetricName.build(name).tagged(dims.map(d => d.name -> d.label).toMap.asJava)
     }
   }